Add multi-threading to the server
