+++ specs/alexandrepiveteau-glsubri/code/src/main/java/protocol/Protocol.java
@@ -1,0 +1,16 @@
+package protocol;
+
+/**
+ * Declarations and constants that are relevant to both ends of the protocol.
+ *
+ * @author Alexandre Piveteau
+ * @author Guy-Laurent Subri
+ */
+public class Protocol {
+  private Protocol() {}
+
+  /**
+   * The port that is defined in the protocol spec.
+   */
+  public static final int HOST_PORT = 8080;
+}
