+++ specs/alexandrepiveteau-glsubri/code/src/main/java/server/Operation.java
@@ -14,6 +14,29 @@ public enum Operation {
     this.behavior = function;
   }
 
+  /**
+   * Returns the {@link Operation} corresponding to a certain String message. Messages should be
+   * encoded to UTF-8. The messages are considered from the server perspective.
+   *
+   * @param message The message that was received.
+   * @return The corresponding {@link Operation} for the message.
+   * @throws MalformedMessageException If the message is not valid.
+   */
+  public static Operation fromMessage(String message) throws MalformedMessageException {
+    switch (message) {
+      case "O ADD":
+        return Operation.ADD;
+      case "O SUB":
+        return Operation.SUB;
+      case "O MUL":
+        return Operation.MUL;
+      case "O DIV":
+        return Operation.DIV;
+      default:
+        throw new MalformedMessageException(message);
+    }
+  }
+
   public int perform(int a, int b) throws IllegalArgumentException {
     try {
       return this.behavior.apply(a, b);
